+++ vagas/src/app/painel-vagas/painel-vagas.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Vaga } from '../models/Vagas.model';
 import { VagasService } from '../vagas.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-painel-vagas',
@@ -22,11 +23,25 @@ export class PainelVagasComponent implements OnInit {
       vaga => {this.vaga = new Vaga(0, "", "", "", 0)},
       err => {console.log("erro ao cadastrar")}
     );
-      alert("VAGA ATUALIZADA");
+
+ const Toast = Swal.mixin({
+  toast: true,
+  position: 'top-end',
+  showConfirmButton: false,
+  timer: 1000,
+  timerProgressBar: true,
+  didOpen: (toast) => {
+    toast.addEventListener('mouseenter', Swal.stopTimer)
+    toast.addEventListener('mouseleave', Swal.resumeTimer)
+  }
+})
+Toast.fire({
+  icon: 'success',
+  title: 'Signed in successfully'
+}).then(()=> location.href = "/mural")
+
       localStorage.removeItem('id');
       localStorage.clear();
-      
-      window.location.href = "/mural";
   }
 
 }
